Add fuzzing tests and fix bugs found fuzzing
What
===
Add fuzzing tests. Change the decode and recode of the string
logic to preserve the original bytes for parts of the string not
rewritten.

Why
===
When decoding a rune out of a string and reencoding it, there's no
guarantee the byte-by-byte representation of that rune we write will
have matched what we read. This is because there is no definitive
normalized format for unicode strings and there can be several ways to
represent the same unicode code points. When rebuilding the string we
can ensure the existing bytes are preserved by simply writing out the
existing bytes instead of reencoding the rune we read out. This is also
subsequently more performant since we're replacing a reencode with a
simple byte copy.